+++ src/Shipment/Changelog.md
@@ -2,10 +2,9 @@
 
 ## 3.x Series
 
-## Unreleased
-##### 2023-XX-YY
+## 3.6.1
+##### 2023-03-09
 
-- Fixed the `ConfigurableModel` trait to handle json strings and other arrayable fields in the underlying model
 - Fixed incorrect shipping method list due to a bug in v2.5.0 of the Address module (bump to v2.5.1)
 
 ## 3.6.0